frontlib: Renamed cfg to scheme, Un-refcounted some types, small API adjustments
--HG--
rename : project_files/frontlib/model/cfg.c => project_files/frontlib/model/scheme.c
rename : project_files/frontlib/model/cfg.h => project_files/frontlib/model/scheme.h

+++ project_files/frontlib/ipc/gameconn.h
@@ -81,7 +81,7 @@ void flib_gameconn_onChat(flib_gameconn *conn, void (*callback)(void* context, c
  * Called when the game ends
  * handleGameRecorded(void *context, const uint8_t *record, int size, bool isSavegame)
  */
-void flib_gameconn_onGameRecorded(flib_gameconn *conn, void (*callback)(void *context, const uint8_t *record, int size, bool isSavegame), void* context);
+void flib_gameconn_onGameRecorded(flib_gameconn *conn, void (*callback)(void *context, const uint8_t *record, size_t size, bool isSavegame), void* context);
 
 /**
  * Called when the game ends
